% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{add_paras}
\alias{add_paras}
\title{Add parameters to an index table object}
\usage{
add_paras(data, para_tbl, by)
}
\arguments{
\item{data}{a \code{idx_tbl} object}

\item{para_tbl}{a tibble or data frame object with parameter of variables}

\item{by}{a single column name (support tidyselect) in the `para_tbl` that
maps to the variable name in the data}
}
\value{
an index object
}
\description{
The function joins the parameter table to the `paras` element of an index
table object.
}
\examples{
init(gggi) |> add_paras(gggi_weights, by = "variable")
}
