% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aqi_ref_tbl}
\alias{aqi_ref_tbl}
\alias{pollutant_ref_tbl}
\alias{aqi}
\title{Air Quality Index (AQI)}
\format{
The aqi data contains daily PM2.5 values in Travis county, Austin, Texas,
USA in 2024, measured in three monitor sites. The data is a
tibble with 272 rows and 9 variables:

\describe{
\item{pollutant}{name of pollutant (PM2.5)}
\item{code}{a five-digit code assigned to each pollutant}
\item{date}{date of measurement}
\item{value}{the measured value of PM2.5}
\item{aqi}{the calculated API value}
\item{long}{longitude of the monitor site}
\item{lat}{latitude of the monitor site}
\item{site_code}{site code}
\item{site_name}{site name}
}

The aqi_ref_tbl and pollutant_ref_tbl data contain the breakpoints for
the AQI and for each of the six pollutants (Ozone, PM2.5, PM10, CO, SO2, NO2).
The aqi_ref_tbl data is a tibble with 5 rows and 3 variables:
\describe{
\item{group}{corresponding group category, from "Good" to "Very Unhealthy"}
\item{low}{the low breakpoint of a certain pollutant group}
\item{high}{the high breakpoint of a certain pollutant group}
}

The pollutant_ref_tbl data is a tibble with 30 rows and 5 variables.
}
\usage{
aqi_ref_tbl

pollutant_ref_tbl

aqi
}
\description{
Data for constructing Air Quality Index (AQI), extracted from the Technical
Assistance Document for the Reporting of Daily Air Quality.
}
\references{
https://document.airnow.gov/technical-assistance-document-for-the-reporting-of-daily-air-quailty.pdf
}
\keyword{datasets}
