% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aqi_table}
\alias{aqi_table}
\title{AQI table}
\format{
A tibble with 3 columns:
\describe{
\item{AQI_low_breakpoint}{the low breakpoint of AQI group}
\item{AQI_high_breakpoint}{the high breakpoint of AQI group}
\item{group}{corresponding group category, from "Good" to "Very Unhealthy"}
}
}
\usage{
aqi_table
}
\description{
A table contains AQI breakpoints as well as corresponding groups, extracted
from the Technical Assistance Document for the Reporting of Daily Air Quality.
}
\references{
https://document.airnow.gov/technical-assistance-document-for-the-reporting-of-daily-air-quailty.pdf
}
\keyword{datasets}
