% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aqi_travis}
\alias{aqi_travis}
\title{AQI travis}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 272 rows and 32 columns.
}
\usage{
aqi_travis
}
\description{
A table contains the PM2.5 values in Travis county, Austin, Texas, USA from
2024/01/01 to 2024/03/31, measured from three monitor sites. Data is collected
via EPA Air Quality System (AQS) API and aqsr package.
}
\references{
https://github.com/kpkeller/aqsr
}
\keyword{datasets}
