% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment}
\alias{augment}
\title{Title}
\usage{
augment(data, id = id, date = date, col = col, gamma_adjust = TRUE)
}
\arguments{
\item{data}{data}

\item{id}{id}

\item{date}{date}

\item{col}{col}

\item{gamma_adjust}{gamma_adjust}
}
\value{
asa
}
\description{
Title
}
\examples{
# add later
}
