% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-pet.R
\name{calc_pet}
\alias{calc_pet}
\title{Calculating PET}
\usage{
calc_pet(data, method, id = id, ..., new_name = ".pet")
}
\arguments{
\item{data}{the data}

\item{method}{method to calculate PET}

\item{id}{the site identifier}

\item{...}{the variable to use}

\item{new_name}{output name}
}
\value{
a data frame with the variable pet calculated
}
\description{
Calculating PET
}
\examples{
if (require(SPEI, quietly = TRUE)){
tenterfield \%>\%
  init(id = id, time = ym, indicators = prcp:tavg) \%>\%
  calc_pet(method = "thornthwaite", id = id, Tave = tavg, lat = -29.0479)
}
}
