% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_index.R
\name{compute_indexes}
\alias{compute_indexes}
\title{Title}
\usage{
compute_indexes(.data, ...)
}
\arguments{
\item{.data}{data}

\item{...}{indexes to compute}
}
\value{
output
}
\description{
Title
}
\examples{
library(lmomco)
library(lubridate)
library(SPEI)
res <- tenterfield \%>\%
init(id = id, time = ym) \%>\%
  compute_indexes(
    spi = idx_spi(.dist = loglogistic()),
    spei = idx_spei(),
    edi = idx_edi()
  )
}
