% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-dim-red.R
\name{dimension_reduction}
\alias{dimension_reduction}
\title{A dimension reduction module}
\usage{
dimension_reduction(data, ...)
}
\arguments{
\item{data}{an `idx_tbl` object}

\item{...}{expression to evaluate}
}
\value{
an `idx_tbl` object
}
\description{
need detailed documentation on the module
}
