% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-dist-fit.R
\name{dist_fit}
\alias{dist_fit}
\title{The distribution fit module}
\usage{
dist_fit(
  .data,
  .dist,
  .gran = "month",
  .var,
  .n_boot = 1,
  .boot_seed = 123,
  .method = "lmoms"
)
}
\arguments{
\item{.data}{an index table object}

\item{.dist}{distributions}

\item{.gran}{granularity}

\item{.var}{variable}

\item{.n_boot}{number of bootstrap samples, default to 1}

\item{.boot_seed}{the seed of bootstrap sampling}

\item{.method}{fitting methods, one of lmoms (L-moment), mle (Maximum Likelihood),
and mom (Method of moment)}
}
\value{
a data frame
}
\description{
The distribution fit module
}
\examples{
# TODO: need to add examples
}
