% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-distributions.R
\name{gamma}
\alias{gamma}
\alias{normal}
\alias{weibull}
\alias{loglogistic}
\alias{lognormal}
\alias{gev}
\alias{pearsonIII}
\title{Distribution functions}
\usage{
gamma()

normal()

weibull()

loglogistic()

lognormal()

gev()

pearsonIII()
}
\value{
a character representation of the distribution
}
\description{
Distribution functions
}
