% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension-reduction.R
\name{dimension_reduction}
\alias{dimension_reduction}
\alias{aggregate_linear}
\alias{aggregate_geometrical}
\alias{manual_input}
\title{The dimension reduction module}
\usage{
dimension_reduction(data, ...)

aggregate_linear(formula, weight)

aggregate_geometrical(formula)

manual_input(formula)
}
\arguments{
\item{data}{an `idx_tbl` object}

\item{...}{expression to evaluate}

\item{formula}{a formula of the dimension reduction expression}

\item{weight}{a variable in the `paras` table of an `idx_tbl` object}
}
\value{
an index table object
}
\description{
The dimension reduction module
}
