% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drought-indexes.R
\name{trans_thornthwaite}
\alias{trans_thornthwaite}
\title{Misc functions used for computing drought indexes}
\usage{
trans_thornthwaite(var, lat, na.rm = FALSE, verbose = TRUE)
}
\arguments{
\item{var}{the variable to be transformed, see [tidyindex::variable_trans()]}
}
\description{
Misc functions used for computing drought indexes
}
