% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indri-package.R
\docType{package}
\name{indri-package}
\alias{indri}
\alias{indri-package}
\title{indri: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/huizezhang-sherry/indri}
  \item Report bugs at \url{https://github.com/huizezhang-sherry/indri/issues}
}

}
\keyword{internal}
