% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollutant_table}
\alias{pollutant_table}
\title{Pollutant table}
\format{
A tibble with 5 columns:
\describe{
\item{pollutant}{name of pollutant}
\item{pollutant_code}{a five-digit code assigned to each pollutant}
\item{low_breakpoint}{the low breakpoint of a certain pollutant group}
\item{high_breakpoint}{the high breakpoint of a certain pollutant group}
\item{group}{corresponding group category, from "Good" to "Very Unhealthy"}
}
}
\usage{
pollutant_table
}
\description{
A table contains breakpoints for six major pollutants: Ozone (O3), PM2.5,
PM10, Carbon Monoxide (CO), Sulfur Dioxide (SO2), Nitrogen Dioxide (NO2).
}
\references{
https://document.airnow.gov/technical-assistance-document-for-the-reporting-of-daily-air-quailty.pdf
}
\keyword{datasets}
