% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-table-methods.R
\name{print.idx_tbl}
\alias{print.idx_tbl}
\title{print methods}
\usage{
\method{print}{idx_tbl}(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{additional argument}
}
\value{
the print
}
\description{
print methods
}
