% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{queensland}
\alias{queensland}
\title{Precipitation in the Queensland stations from 1990 to 2020}
\format{
A tibble with three columns:
\describe{
\item{id}{station ID, corresponds to tenterfied station}
\item{ym}{date in yearmonth}
\item{prcp}{summed monthly precipitation from daily data}
\item{tmax}{max temp}
\item{tmin}{min temp}
\item{tavg}{average temp}
\item{long}{longitude}
\item{lat}{latitude}
\item{name}{name}
}
}
\usage{
queensland
}
\description{
Precipitation in the Queensland stations from 1990 to 2020
}
\keyword{datasets}
