% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-rescaling.R
\name{rescaling}
\alias{rescaling}
\title{Rescaling}
\usage{
rescaling(data, ..., .method = NULL, .vars = NULL, .new_name = NULL)
}
\arguments{
\item{data}{an index table object}

\item{...}{expression}

\item{.method}{the method use}

\item{.vars}{variables}

\item{.new_name}{the new name}
}
\value{
an index table
}
\description{
Specify either as an expression (like mutate) or using .method.
}
\details{
When using .method, specifies the varaibles it applies to (`.vars`), can be multiple, use tidyselect syntax
additional parameters will be parsed from `...`
`.new_names` needs to be of same length as `.vars`
}
\examples{

dt <- hdi \%>\% init(id = country, indicators = life_exp:gni_pc)
# only print out the operation table
show_summary <- function(obj) obj \%>\% .[["op"]]

# single variable:
show_summary(dt \%>\% rescaling(life_exp = rescale_minmax(life_exp)))
show_summary(dt \%>\% rescaling(.method = rescale_minmax, .vars = life_exp, min = 20, max = 85))
show_summary(dt \%>\% rescaling(.method = rescale_minmax, .vars = life_exp, .new_name = "life_exp2"))

# use tidyselect to apply the .method to multiple variables
show_summary(dt \%>\% rescaling(.method = rescale_minmax, .vars = life_exp:exp_sch))
# can also supply additional parameters
show_summary(dt \%>\% rescaling(.method = rescale_minmax,
                 .vars = life_exp:exp_sch,
                 min = c(20, 0), max = c(85, 18),
                 .new_name = c("life_exp2", "exp_sch")))
}
\seealso{
[rescale_zscore()], [rescale_minmax()], [rescale_center()]
}
