% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-swap-values.R
\name{swap_values}
\alias{swap_values}
\title{Title}
\usage{
swap_values(obj, .id, .param, .values, .raw_data)
}
\arguments{
\item{obj}{obj}

\item{.id}{module}

\item{.param}{the parameter to swap}

\item{.values}{values}

\item{.raw_data}{the original data}
}
\value{
an index table
}
\description{
Title
}
