% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch-values.R
\name{switch_values}
\alias{switch_values}
\title{Title}
\usage{
switch_values(obj, module, step, res, var, values, raw_data)
}
\arguments{
\item{obj}{obj}

\item{module}{module}

\item{step}{step}

\item{res}{res}

\item{var}{var}

\item{values}{values}

\item{raw_data}{the original data}
}
\value{
an indri object
}
\description{
Title
}
