% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal-aggregation.R
\name{temporal_aggregate}
\alias{temporal_aggregate}
\alias{temporal_rolling_window}
\title{The temporal processing module}
\usage{
temporal_aggregate(data, ...)

temporal_rolling_window(
  var,
  scale,
  .before = 0L,
  .step = 1L,
  .complete = FALSE,
  ...
)
}
\arguments{
\item{data}{an index table object, see [tidyindex::init()]}

\item{...}{an temporal processing object of class \code{temporal_agg}}
}
\value{
an index table object
}
\description{
The temporal processing module is used to aggregate data along the temporal
dimension. Current available aggregation recipe includes
 \code{temporal_rolling_window}.
}
\examples{
tenterfield |>
  init() |>
  temporal_aggregate(.agg = temporal_rolling_window(prcp, scale = 12))

# with multiple scales
tenterfield |>
  init() |>
  temporal_aggregate(temporal_rolling_window(prcp, scale = c(12, 24)))
}
