% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tenterfield}
\alias{tenterfield}
\title{Precipitation in the Tenterfield station from 1990 to 2020}
\format{
A tibble with three columns:
\describe{
\item{id}{station ID, corresponds to tenterfied station}
\item{ym}{date in yearmonth}
\item{prcp}{summed monthly precipitation from daily data}
}
}
\usage{
tenterfield
}
\description{
Precipitation in the Tenterfield station from 1990 to 2020
}
\keyword{datasets}
