% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{theme_benchmark}
\alias{theme_benchmark}
\title{Title}
\usage{
theme_benchmark(yintercept = -2, linetype = "dashed")
}
\arguments{
\item{yintercept}{intercpt}

\item{linetype}{linetype}
}
\value{
a ggplot2 object
}
\description{
Title
}
\examples{
if (require("ggplot2", quietly = TRUE) ){
dplyr::tibble(x = 1:100, y = rnorm(100, sd = 2)) \%>\%
  ggplot(aes(x = x, y =y )) +
  geom_line() +
  theme_benchmark()
}
}
