% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{thornthwaite}
\alias{thornthwaite}
\title{The thornthwaite method in calculating SPEI}
\usage{
thornthwaite(Tave, lat, ...)
}
\arguments{
\item{Tave}{tavg}

\item{lat}{lat}

\item{...}{other arguement}
}
\value{
a data frame
}
\description{
The thornthwaite method in calculating SPEI
}
\examples{
# tobefilled
}
