% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-methods.R
\name{tidy.indri}
\alias{tidy.indri}
\title{Title}
\usage{
\method{tidy}{indri}(.data, ...)
}
\arguments{
\item{.data}{data}

\item{...}{others}
}
\value{
a tidied indri
}
\description{
Title
}
\examples{
# tobefilled
}
