% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indri-package.R
\docType{package}
\name{tidyindex-package}
\alias{tidyindex}
\alias{tidyindex-package}
\title{tidyindex: A general data pipeline for computing indexes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to construct and analyse indexes in a pipeline workflow. `tidyindex` contains modules for transforming variables, aggregating variables across time, reducing data dimension through weighting, and fitting distributions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/huizezhang-sherry/indri}
  \item Report bugs at \url{https://github.com/huizezhang-sherry/indri/issues}
}

}
\keyword{internal}
