% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var-trans.R
\name{var_trans}
\alias{var_trans}
\title{Variable transformation}
\usage{
var_trans(data, method = NULL, vars, ..., param_tbl, new_name = NULL)
}
\arguments{
\item{data}{data}

\item{method}{the method use}

\item{vars}{variables}

\item{...}{expression}

\item{param_tbl}{parameter table}

\item{new_name}{the new name}
}
\value{
an indri object
}
\description{
Variable transformation
}
