% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tenterfield}
\alias{tenterfield}
\alias{aus_climate}
\alias{queensland}
\title{Weather data for in-situ stations in Queensland from 1990 to 2020}
\format{
A tibble with 9 columns:
\describe{
\item{id}{station ID, ASN000xxxxx}
\item{ym}{date in `tsibble::yearmonth` format}
\item{prcp}{aggregated monthly precipitation from daily data}
\item{tmax, tmin, tavg}{maximum/minimum/ average temperature}
\item{long, lat}{longitude and latitude of the station}
\item{name}{station name}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 52373 rows and 9 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11252 rows and 9 columns.
}
\usage{
tenterfield

aus_climate

queensland
}
\description{
Weather data for in-situ stations in Queensland from 1990 to 2020
}
\keyword{datasets}
