% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate across time}
\usage{
aggregate(data, var, scale, ..., id = NULL, index = NULL, na.rm = TRUE)
}
\arguments{
\item{data}{a data frame}

\item{var}{the variable to aggregate along}

\item{scale}{the time scale to aggregate}

\item{...}{ignore}

\item{id}{the site identifier}

\item{index}{the temporal identifier}

\item{na.rm}{logical, whether to remove the pending NAs when aggregated}
}
\value{
a data frame with the variable specified aggregated in scale
}
\description{
Aggregate across time
}
\examples{
# first 11 rows should be NA since aggregate by 12 months
(res <- tenterfield \%>\% aggregate(var = prcp, scale = 12))
res \%>\% dplyr::filter(!is.na(.agg))

# with multiple scales
tenterfield \%>\% aggregate(var = prcp, scale = c(12, 24))
}
