% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-augment.R
\name{augment}
\alias{augment}
\title{Title}
\usage{
augment(.data, .var = var, .gamma_adjust = TRUE, .new_name = ".index")
}
\arguments{
\item{.data}{data}

\item{.var}{col}

\item{.gamma_adjust}{gamma_adjust}

\item{.new_name}{new name}
}
\value{
asa
}
\description{
Title
}
\examples{
# add later
}
