% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-augment.R
\name{augment}
\alias{augment}
\title{The normalising module}
\usage{
augment(.data, .var = var, .new_name = ".index")
}
\arguments{
\item{.data}{an index table object}

\item{.var}{the variable to normalise}

\item{.new_name}{character, the new variable name after `augment`}
}
\value{
an index table object
}
\description{
The normalising module
}
\examples{
library(lubridate)
library(lmomco)
tenterfield \%>\%
  init(id = id, time = ym, indicators = prcp:tavg) \%>\%
  aggregate(.var = prcp, .scale = 12) \%>\%
  dist_fit(.dist = gamma(), .method = "lmoms", .var = .agg) \%>\%
  augment(.var = .agg)
}
