% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_index.R
\name{compute_indexes}
\alias{compute_indexes}
\title{Title}
\usage{
compute_indexes(.data, .index_value = TRUE, ...)
}
\arguments{
\item{.data}{data}

\item{.index_value}{logical; whether to output index values}

\item{...}{indexes to compute}
}
\value{
output
}
\description{
Title
}
\examples{
library(lmomco)
library(lubridate)
library(SPEI)
res <- tenterfield \%>\%
 init(id = id, time = ym) \%>\%
 compute_indexes(
    spi = idx_spi(.dist = loglogistic()),
    spei = idx_spei(),
    edi = idx_edi()
 )
res \%>\%
 ggplot(aes(x = ym, y = .index, color = .idx)) +
 geom_line() +
 theme_benchmark()
}
