% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-indexes.R
\name{compute_indexes}
\alias{compute_indexes}
\title{Calculate multiple indexes at once}
\usage{
compute_indexes(.data, .index_value = TRUE, ...)
}
\arguments{
\item{.data}{an `idx_tbl` object}

\item{.index_value}{logical; whether to output index values}

\item{...}{indexes to compute}
}
\value{
output
}
\description{
Calculate multiple indexes at once
}
\examples{
\dontrun{
library(lmomco)
library(lubridate)
library(tsibble)
library(ggplot2)
res <- tenterfield |>
 init(id = id, time = ym) |>
 compute_indexes(
    spi = idx_spi(.dist = loglogistic()),
    spei = idx_spei(.pet_method = "thornthwaite" ,.tavg = tavg, .lat = lat),
    edi = idx_edi()
 )
res |>
 ggplot(aes(x = ym, y = .index, color = .idx)) +
 geom_line() +
 theme_benchmark()
}
}
