% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim-red.R
\name{dim_red}
\alias{dim_red}
\title{Dimension reduction}
\usage{
dim_red(data, expr, new_name)
}
\arguments{
\item{data}{data}

\item{expr}{expression to evaluate}

\item{new_name}{the new name}
}
\value{
a list object
}
\description{
Dimension reduction
}
