% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{dist_fit}
\alias{dist_fit}
\title{Title}
\usage{
dist_fit(
  .data,
  .dist,
  .gran = "month",
  .var,
  .n_boot = 1,
  .boot_seed = 123,
  .method = "lmoms",
  .gamma_adjust = TRUE
)
}
\arguments{
\item{.data}{a data frame}

\item{.dist}{distributions}

\item{.gran}{granularity}

\item{.var}{variable}

\item{.n_boot}{number of bootstrap samples, default to 1}

\item{.boot_seed}{the seed of bootstrap sampling}

\item{.method}{fitting methods, one of lmoms (L-moment), mle (Maximum Likelihood),
and mom (Method of moment)}

\item{.gamma_adjust}{adjustment for gamma, add reference}
}
\value{
a data frame
}
\description{
Title
}
\examples{
# TODO: need to add examples
}
