% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hdi}
\alias{hdi}
\alias{hdi_scales}
\title{Human Development Index (2022)}
\format{
A tibble with three columns:
\describe{
\item{id}{the row number}
\item{country}{191 countries with computed HDI}
\item{hdi}{the HDI index value}
\item{life_exp}{life expectancy}
\item{exp_sch}{expected schooling}
\item{avg_sch}{average schooling}
\item{gni_pc}{GNI per capital, logged}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 5 columns.
}
\usage{
hdi

hdi_scales
}
\description{
Human Development Index (2022)
}
\references{
https://hdr.undp.org/data-center/human-development-index#/indicies/HDI
}
\keyword{datasets}
