% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructed-idx.R
\name{idx_spei}
\alias{idx_spei}
\alias{idx_spi}
\alias{idx_rdi}
\alias{idx_edi}
\title{Pre-built indexes}
\usage{
idx_spei(
  data,
  id,
  time,
  pet_method = "thornthwaite",
  .scale = 12,
  dist = loglogistic()
)

idx_spi(data, id, time, dist = gamma(), .scale = 12)

idx_rdi(data, id, time, pet_method = "thornthwaite", .scale = .scale)

idx_edi(data, id, time, .scale = .scale)
}
\arguments{
\item{data}{data}

\item{id}{id}

\item{time}{time}

\item{pet_method}{PET method}

\item{.scale}{scale}

\item{dist}{distribution}
}
\value{
calculated indexes
}
\description{
Pre-built indexes
}
\examples{
library(lmomco)
library(lubridate)
library(SPEI)
tenterfield \%>\% idx_spi(id = id, time = ym, dist = list(gev(), loglogistic()))
tenterfield \%>\% idx_spei(id = id, time = ym)
tenterfield \%>\% idx_edi(id = id, time = ym, .scale = 12)
}
