% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructed-idx.R
\name{idx_spei}
\alias{idx_spei}
\alias{idx_spi}
\alias{idx_rdi}
\alias{idx_edi}
\title{Pre-built indexes}
\usage{
idx_spei(
  data,
  id,
  time,
  .pet_method = "thornthwaite",
  .tavg,
  .lat,
  .scale = 12,
  .dist = loglogistic(),
  .new_name = ".index"
)

idx_spi(data, id, time, .dist = gamma(), .scale = 12, .new_name = ".index")

idx_rdi(
  data,
  id,
  time,
  .pet_method = "thornthwaite",
  .scale = .scale,
  .new_name = ".index"
)

idx_edi(data, id, time, .scale = 12, .new_name = ".index")
}
\arguments{
\item{data}{data}

\item{id}{id}

\item{time}{time}

\item{.pet_method}{PET method}

\item{.tavg}{average temperature}

\item{.lat}{latitude}

\item{.scale}{scale}

\item{.dist}{distribution}

\item{.new_name}{new names}
}
\value{
calculated indexes
}
\description{
Pre-built indexes
}
\examples{
library(lmomco)
library(lubridate)
tenterfield \%>\% init(id = id, time = ym) \%>\%
   idx_spei(.tavg = tavg, .lat = lat)
tenterfield \%>\% init(id = id, time = ym) \%>\% idx_spi()
tenterfield \%>\% init(id = id, time = ym) \%>\% idx_edi()
}
