% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-init.R
\name{init}
\alias{init}
\title{Initialise the pipeline}
\usage{
init(data, ...)
}
\arguments{
\item{data}{the dataset}

\item{...}{additional argument}
}
\value{
a list object
}
\description{
Initialise the pipeline
}
