% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\alias{print.idx_tbl}
\title{Initialise the tidyindex pipeline}
\usage{
init(data)

\method{print}{idx_tbl}(x)
}
\arguments{
\item{data}{a tibble or data frame to be converted into a index object}

\item{x}{an index object}
}
\value{
an index table object
}
\description{
Initialise an index table object with a data frame or a tibble.
}
\examples{
init(hdi)
init(gggi)
}
