% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-rescale.R
\name{rescale_zscore}
\alias{rescale_zscore}
\alias{rescale_minmax}
\alias{rescale_center}
\title{Rescaling functions}
\usage{
rescale_zscore(var, na.rm = TRUE)

rescale_minmax(var, min = NULL, max = NULL, na.rm = TRUE, censor = TRUE)

rescale_center(var, na.rm = TRUE)
}
\arguments{
\item{var}{variable}

\item{na.rm}{logical; whether to remove NAs}

\item{min}{min}

\item{max}{max}

\item{censor}{logical; whether to censor points outside min and max, if provided}
}
\value{
a vector of rescaled variable
}
\description{
guide on implementing new rescaling functions: a vector function that
takes a single variable, var, with additional parameters
}
\details{
A known problem: `dt %>% rescaling(life_exp = rescale_minmax, var = life_exp)` will fail
}
\seealso{
[rescaling()]
}
