% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-rescaling.R
\name{rescaling}
\alias{rescaling}
\alias{rescale_zscore}
\alias{rescale_minmax}
\alias{rescale_center}
\title{Rescaling}
\usage{
rescaling(data, ..., .method = NULL, .vars = NULL, .new_name = NULL)

rescale_zscore(var, na.rm = TRUE)

rescale_minmax(var, min = NULL, max = NULL, na.rm = TRUE, censor = TRUE)

rescale_center(var, na.rm = TRUE)
}
\arguments{
\item{data}{an index table object}

\item{...}{expression}

\item{.method}{the method use for rescaling, currently rescale_minmax, rescale_zscore, rescale_center}

\item{.vars}{variables to rescale}

\item{.new_name}{the new name for the rescaled variable, optional}

\item{var}{variable, argument in the rescale_* functions}

\item{na.rm}{logical; whether to remove NAs,  argument in the rescale_* functions}

\item{min, max}{the minimum and maximum value used in rescale_minmax}

\item{censor}{logical; whether to censor points outside min and max, if provided}
}
\description{
Specify either as an expression (like mutate) or using .method.
}
\details{
When using .method, specifies the varaibles it applies to (`.vars`), can be multiple, use tidyselect syntax
additional parameters will be parsed from `...`
`.new_names` needs to be of same length as `.vars`
}
\examples{

dt <- hdi \%>\% init(id = country, indicators = life_exp:gni_pc)
# only print out the operation table
show_summary <- function(obj) obj \%>\% .[["op"]]

# single variable:
show_summary(dt \%>\% rescaling(life_exp = rescale_minmax(life_exp)))
show_summary(dt \%>\% rescaling(.method = rescale_minmax, .vars = life_exp, min = 20, max = 85))
show_summary(dt \%>\% rescaling(.method = rescale_minmax, .vars = life_exp, .new_name = "life_exp2"))

# use tidyselect to apply the .method to multiple variables
show_summary(dt \%>\% rescaling(.method = rescale_minmax, .vars = life_exp:exp_sch))
# can also supply additional parameters
show_summary(dt \%>\% rescaling(.method = rescale_minmax,
                 .vars = life_exp:exp_sch,
                 min = c(20, 0), max = c(85, 18),
                 .new_name = c("life_exp2", "exp_sch")))
}
