% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaling.R
\name{rescaling}
\alias{rescaling}
\alias{rescale_zscore}
\alias{rescale_minmax}
\alias{rescale_center}
\title{The rescaling module}
\usage{
rescaling(data, ...)

rescale_zscore(var, na.rm = TRUE)

rescale_minmax(var, min = NULL, max = NULL, na.rm = TRUE, censor = TRUE)

rescale_center(var, na.rm = TRUE)
}
\arguments{
\item{data}{an index table object, see [tidyindex::init()]}

\item{...}{expression, in the syntax of \code{NEW_NAME = rescale_*(...)}}

\item{var}{the variable(s) to rescale, accept tidyselect syntax}

\item{na.rm}{used in \code{rescale_*()}; logical; whether to remove NAs}

\item{min, max}{used in \code{rescale_minmax()}, the minimum and maximum value}

\item{censor}{logical; whether to censor points outside min and max, if provided}
}
\description{
The rescale module accepts a [dplyr::mutate()] alike expression to change
the scale of the variable(s). Currently avaiable rescaling: []
}
\examples{
dt <- hdi \%>\% init()
dt |> rescaling(life_exp = rescale_zscore(life_exp))
dt |> rescaling(life_exp2 = rescale_minmax(life_exp, min = 20, max = 85))
dt |> rescaling(life_exp = rescale_minmax(life_exp, min = 20, max = 85))
}
