% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-swap-exprs.R, R/analysis-swap-values.R
\name{swap_exprs}
\alias{swap_exprs}
\alias{swap_values}
\title{Replace a value/expression in the pipeline with other alternatives}
\usage{
swap_exprs(obj, .var, .exprs, .raw_data)

swap_values(obj, .id, .param, .values, .raw_data)
}
\arguments{
\item{obj}{an index table object}

\item{.raw_data}{the initial index table created}

\item{.id, .var}{the ID/ variable used to locate the operation of the value/ expression}

\item{.param}{the parameter to swap}

\item{.values, .exprs}{the values or expressions to test}
}
\value{
an index table
}
\description{
Replace a value/expression in the pipeline with other alternatives
}
