% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-swap-exprs.R
\name{swap_exprs}
\alias{swap_exprs}
\title{Title}
\usage{
swap_exprs(obj, .var, .exprs, .raw_data)
}
\arguments{
\item{obj}{an index table}

\item{.var}{the variable to try alternative expression}

\item{.exprs}{the new expressions to test}

\item{.raw_data}{the initial index table created}
}
\value{
an index table
}
\description{
Title
}
