% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{thornthwaite}
\alias{thornthwaite}
\title{Title}
\usage{
thornthwaite(.data, .tavg, .lat)
}
\arguments{
\item{.data}{data}

\item{.tavg}{tavg}

\item{.lat}{lat}
}
\value{
a data frame
}
\description{
Title
}
\examples{
# tobefilled
}
