% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-table-methods.R
\name{tidy.idx_tbl}
\alias{tidy.idx_tbl}
\title{A tidy method for the index table}
\usage{
\method{tidy}{idx_tbl}(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{others}
}
\value{
a tidied index table
}
\description{
A tidy method for the index table
}
\examples{
# tobefilled
}
