% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyindex-package.R
\docType{package}
\name{tidyindex-package}
\alias{tidyindex}
\alias{tidyindex-package}
\title{tidyindex: A Tidy Data Pipeline to Construct, Compare, and Analyse Indexes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Construct and analyse indexes in a pipeline tidy workflow. 'tidyindex' contains modules for transforming variables, aggregating variables across time, reducing data dimension through weighting, and fitting distributions. A manuscript describing the methodology can be found at \url{https://github.com/huizezhang-sherry/paper-tidyindex}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/huizezhang-sherry/tidyindex}
  \item Report bugs at \url{https://github.com/huizezhang-sherry/tidyindex/issues}
}

}
\author{
\strong{Maintainer}: H. Sherry Zhang \email{huize.zhang@monash.edu} (\href{https://orcid.org/0000-0002-7122-1463}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Dianne Cook \email{dicook@monash.edu} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID})
  \item Ursula Laa \email{ursula.laa@boku.ac.at} (\href{https://orcid.org/0000-0002-0249-6439}{ORCID})
  \item Nicolas Langrené \email{nicolas.langrene@csiro.au} (\href{https://orcid.org/0000-0001-7601-4618}{ORCID})
  \item Patricia Menéndez \email{patricia.menendez@monash.edu} (\href{https://orcid.org/0000-0003-0701-6315}{ORCID})
}

}
\keyword{internal}
