% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-var-trans.R
\name{var_trans}
\alias{var_trans}
\title{Variable transformation}
\usage{
var_trans(data, ..., .method = NULL, .vars = NULL, .new_name = NULL)
}
\arguments{
\item{data}{an index table object}

\item{...}{expression}

\item{.method}{the method use}

\item{.vars}{variables}

\item{.new_name}{the new name}
}
\value{
an index table
}
\description{
Specify either as an expression (like mutate) or using .method.
}
\details{
When using .method, specifies the varaibles it applies to (`.vars`), can be multiple, use tidyselect syntax
additional parameters will be parsed from `...`
`.new_names` needs to be of same length as `.vars`
}
\examples{

dt <- hdi \%>\% init(id = country, indicators = life_exp:gni_pc)
# single variable:
dt \%>\% var_trans(life_exp = rescale_minmax(life_exp))
dt \%>\% var_trans(.method = rescale_minmax, .vars = life_exp, min = 20, max = 85)
dt \%>\% var_trans(.method = rescale_minmax, .vars = life_exp, .new_name = "life_exp2")

# use tidyselect to apply the .method to multiple variables
dt \%>\% var_trans(.method = rescale_minmax, .vars = life_exp:exp_sch)
# can also supply additional parameters
dt \%>\% var_trans(.method = rescale_minmax,
                 .vars = life_exp:exp_sch,
                 min = c(20, 0), max = c(85, 18),
                 .new_name = c("life_exp2", "exp_sch"))
}
