% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-var-trans.R
\name{var_trans}
\alias{var_trans}
\title{Variable transformation}
\usage{
var_trans(data, ..., .method = NULL, .vars = NULL, .new_name = NULL)
}
\arguments{
\item{data}{an index table object}

\item{...}{expression}

\item{.method}{the method use}

\item{.vars}{variables}

\item{.new_name}{the new name}
}
\value{
an index table
}
\description{
Variable transformation
}
\examples{
#tobefilled
}
