% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-var-trans.R
\name{var_trans}
\alias{var_trans}
\title{The variable transformation module}
\usage{
var_trans(data, ..., .method = NULL, .vars = NULL, .new_name = NULL)
}
\arguments{
\item{data}{an index table object}

\item{...}{the transformation expression}

\item{.method}{the function used for transformation}

\item{.vars}{the variables to transform}

\item{.new_name}{character, the new variable name after var_trans}
}
\value{
an index table
}
\description{
The variable transformation module
}
\examples{
tenterfield |>
  init(id = id, time = ym, indicators = prcp:tavg) |>
  var_trans(.method = thornthwaite, .vars = tavg, lat = lat, .new_name = "pet")
}
